requirement.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.50.0
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
scikit-fuzzy==0.4.2
scipy==1.12.0
six==1.16.0
